+++ Editor/AnimationExpressEditorUtils.cs
@@ -1,0 +1,10 @@
+namespace AnimExpressEditor
+{
+	/// <summary>
+	/// Class which contains shared information across editor classes.
+	/// </summary>
+	public static class AnimationExpressEditorUtils
+	{
+		public const float COPY_BUTTON_WIDTH = 30f;
+	}
+}