+++ Editor/FramePropertyDrawer.cs
@@ -20,11 +20,22 @@ namespace AnimExpressEditor
 			var duration = property.FindPropertyRelative("duration");
 
 			Rect spriteRect = position;
-			spriteRect.width = position.width - DURATION_FIELD_WIDTH;
+			spriteRect.width = position.width - AnimationExpressEditorUtils.COPY_BUTTON_WIDTH - DURATION_FIELD_WIDTH;
+
+			Rect copyButtonRect = position;
+			copyButtonRect.width = AnimationExpressEditorUtils.COPY_BUTTON_WIDTH;
+			copyButtonRect.x += spriteRect.width;
+
 			Rect durationRect = position;
 			durationRect.width = DURATION_FIELD_WIDTH;
-			durationRect.x += spriteRect.width;
+			durationRect.x += spriteRect.width + AnimationExpressEditorUtils.COPY_BUTTON_WIDTH;
+
 			EditorGUI.PropertyField(spriteRect, sprite, GUIContent.none);
+			if (GUI.Button(copyButtonRect, EditorGUIUtility.IconContent("Clipboard")))
+			{
+				GUIUtility.systemCopyBuffer = $"\"{sprite.objectReferenceValue.name}\"";
+			}
+
 			EditorGUI.PropertyField(durationRect, duration, GUIContent.none);
 		}
 	}